chore: remove remaining underscore packages
