+++ instrumentation/sources/gingonic/orchestrion.yml
@@ -5,7 +5,7 @@
 ---
 # yaml-language-server: $schema=https://datadoghq.dev/orchestrion/schema.json
 meta:
-  name: github.com/AikidoSec/firewall-go/instrumentation/sources/gin-gonic
+  name: github.com/AikidoSec/firewall-go/instrumentation/sources/gingonic
   description: Gin is a web framework written in Go.
 
 aspects:
@@ -18,7 +18,7 @@ aspects:
       - wrap-expression:
           imports:
             gin: github.com/gin-gonic/gin
-            zen_gin: github.com/AikidoSec/firewall-go/instrumentation/sources/gin-gonic
+            zen_gin: github.com/AikidoSec/firewall-go/instrumentation/sources/gingonic
           template: |-
             func() *gin.Engine {
               e := {{ . }}
